Refactor TestWordAlignedLattice to outside of WordAlignLatticeLexicon in word-align-lattice-lexicon.cc
When it was there, we had to do a weird dynamic_cast<> to check if the
TransitionInformation was a TransitionModel.
